fix #896, merge utilities/custom_logging.py in common/custom_logging.py, enable LOGGER in top level
